requirements.txt
pandas
ta
matplotlib
aiogram
tinkoff-investments
xgboost
pytz
